backend/was/requirements.txt
# backend
bcrypt==3.2.2
Django==4.0.6
djangorestframework==3.13.1
PyMySQL==1.0.2
cryptography==37.0.4
django-cors-headers==3.13.0
drf-yasg==1.21.0
django-environ==0.9.0
django-storages==1.12.3
mysqlclient==2.1.1
elasticsearch==7.17.4
boto3==1.24.33

PyJWT
gunicorn
pandas